Deprecated: image data saved in MySQL

+++ src/controllers/deleteImage.ctrl.js
@@ -1,45 +1,45 @@
+const { env } = require("../config/env");
 const { cloudinary } = require("../config");
 const { connect, disconnect } = require("../database/dbconnection");
 const imagesRepository = require("../repository/images.repo");
 
 const deleteImage = async (req, res) => {
   const { public_id } = req.body;
-  let connection;
+  let connection = null;
 
   try {
-    connection = await connect();
+    if (env.ENABLE_SAVE_MYSQL) connection = await connect();
     if (!public_id) {
-      disconnect(connection);
-      res.status(400).json({ status: "Error", message: "public_id required" });
+      if (env.ENABLE_SAVE_MYSQL) disconnect(connection);
+      return res.status(400).json({ status: "Error", message: "public_id required" });
     }
   } catch (err) {
     console.error(err);
-    disconnect(connection);
-    res.status(500).json({ status: "Error", message: "Internal server error" });
+    if (env.ENABLE_SAVE_MYSQL) disconnect(connection);
+    return res.status(500).json({ data: null, success: false, message: "Internal server error", error: err });
   }
 
-  imagesRepository
-    .deleteImages(connection, public_id)
-    .then(async () => {
+  // #### deprecated ####
+  if (env.ENABLE_SAVE_MYSQL) {
+    imagesRepository
+      .deleteImages(connection, public_id)
+      .then(async () => console.log("Images deleted successfully"))
+      .catch(() => console.error(err))
+      .finally(() => disconnect(connection));
+  }
+  // ####################################
 
-      try {
-        await cloudinary.uploader.destroy(public_id, {
-          folder: process.env.CLOUDINARY_FOLDER_IMAGES,
-        });
-        res.json({ status: "Ok", message: "Image deleted successfully" });
-      } catch (error) {
-        res
-          .status(500)
-          .json({ status: "Error", message: "Internal server error" });
-      }
+  try {
+    await cloudinary.uploader.destroy(public_id, {
+      folder: env.CLOUDINARY_FOLDER_IMAGES,
+    });
+  } catch (err) {
+    return res
+      .status(500)
+      .json({ data: null, success: false, message: "Internal server error", error: err });
+  }
 
-    })
-    .catch(() => {
-      res
-        .status(500)
-        .json({ status: "Error", message: "Internal server error" });
-    })
-    .finally(() => disconnect(connection));
+  return res.status(200).json({ data: null, success: true, message: "Images deleted successfully", error: null });
 };
 
 module.exports = deleteImage;
